Rename json/ -> frame/
In the past it was thought that we would use more than one serialization method
for BAM messages.  Now it is thought that we will only use JSON.  We can now
remove the directory `bam/src/json`.  Instead of just removing one layer of the
directory tree if we rename to `frame` then code will read nicely
e.g. `bam::frame::Request`.
